Save the trace and tick for the player's last collision
Gives information about the last collision the player had, which will be used for treating edgebugs. It does not save collision of straight walls though. The reason for rasing the trace positions for ceiling collisions is that it would not be possible to deduce the ceiling position otherwise, since the ceiling offset is dependent on the collision box height.